rename test_jinja_globals.py & add tests
